+++ src/gnome/gronor_master.F90
@@ -60,8 +60,8 @@ subroutine gronor_master()
   real (kind=8) :: tbuf(18)
   integer :: gtid, lfnmpi
 
-  integer (kind=4) :: ierr,ireq2,ireq9,iremote,ncount,mpitag
-  integer (kind=4) :: status(MPI_STATUS_SIZE)
+  integer :: ierr, ireq2, ireq9, iremote, ncount, mpitag
+  integer :: status(MPI_STATUS_SIZE)
   integer :: err_len, ierr_abort
   character(len=MPI_MAX_ERROR_STRING) :: err_msg
